kvstore/requirements.txt
antiorm==1.2.1
blinker==1.8.2
cffi==1.16.0
click==8.1.7
cryptography==41.0.4
db==0.1.1
db-sqlite3==0.0.1
distlib==0.3.7
distro==1.8.0
dominate==2.9.1
filelock==3.12.4
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
iniparse==0.5
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pathlib2==2.3.7.post1
platformdirs==3.11.0
pycparser==2.21
PyJWT==2.9.0
PyMySQL==1.1.0
pyOpenSSL==23.2.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.32
SQLAlchemy-Utils==0.41.2
typing_extensions==4.12.2
virtualenv==20.24.5
visitor==0.1.3
Werkzeug==3.0.3
WTForms==3.1.2
